[ADF-4482] Refactor Social Component tests
