+++ e2e/pages/adf/demo-shell/socialPage.ts
@@ -15,6 +15,22 @@
  * limitations under the License.
  */
 
-export * from './like.page';
-export * from './rate.page';
-export * from './document-list.page';
+import { by, element } from 'protractor';
+import { BrowserVisibility } from '@alfresco/adf-testing';
+
+export class SocialPage {
+
+    nodeIdField = element(by.css(`input[id="nodeId"]`));
+
+    getNodeIdFieldValue() {
+        BrowserVisibility.waitUntilElementIsVisible(this.nodeIdField);
+        return this.nodeIdField.getAttribute('value');
+    }
+
+    writeCustomNodeId(nodeId: string) {
+        BrowserVisibility.waitUntilElementIsVisible(this.nodeIdField);
+        this.nodeIdField.clear();
+        this.nodeIdField.sendKeys(nodeId);
+    }
+
+}
