+++ lib/testing/src/lib/content-services/pages/like.page.ts
@@ -20,21 +20,9 @@ import { BrowserVisibility } from '../../core/utils/browser-visibility';
 
 export class LikePage {
 
-    nodeIdField = element(by.css(`input[id="nodeId"]`));
     likeCounter = element(by.css(`div[id="adf-like-counter"]`));
     likeButton = element(by.css(`span[class="adf-like-grey"]`));
     unlikeButton = element(by.css(`span[class="adf-like-select"]`));
-
-    getNodeIdFieldValue() {
-        BrowserVisibility.waitUntilElementIsVisible(this.nodeIdField);
-        return this.nodeIdField.getAttribute('value');
-    }
-
-    writeCustomNodeId(nodeId: string) {
-        BrowserVisibility.waitUntilElementIsVisible(this.nodeIdField);
-        this.nodeIdField.clear();
-        this.nodeIdField.sendKeys(nodeId);
-    }
 
     getLikeCounter() {
         BrowserVisibility.waitUntilElementIsVisible(this.likeCounter);
@@ -53,22 +41,6 @@ export class LikePage {
         BrowserVisibility.waitUntilElementIsVisible(this.likeButton);
     }
 
-    rateComponent(rateValue: number) {
-        const unratedStar = element(by.css(`span[id="adf-rate-${rateValue}"]`));
-        BrowserVisibility.waitUntilElementIsClickable(unratedStar);
-        unratedStar.click();
-    }
-
-    isStarRated(rateValue: number) {
-        const ratedStar = element(by.css(`mat-icon[id="adf-colored-star-${rateValue}"]`));
-        return BrowserVisibility.waitUntilElementIsVisible(ratedStar);
-    }
-
-    isNotStarRated(rateValue: number) {
-        const unratedStar = element(by.css(`mat-icon[id="adf-grey-star-${rateValue}"]`));
-        return BrowserVisibility.waitUntilElementIsVisible(unratedStar);
-    }
-
     removeHoverFromLikeButton() {
         browser.actions().mouseMove({x: 200, y: 200}).click().perform();
     }
@@ -81,17 +53,5 @@ export class LikePage {
     getUnlikedIconColor() {
         BrowserVisibility.waitUntilElementIsNotVisible(this.unlikeButton);
         return this.likeButton.getWebElement().getCssValue('color');
-    }
-
-    getRatedStarColor(rateValue: number) {
-        const ratedStar = element(by.css(`mat-icon[id="adf-colored-star-${rateValue}"]`));
-        BrowserVisibility.waitUntilElementIsVisible(ratedStar);
-        return ratedStar.getWebElement().getCssValue('color');
-    }
-
-    getUnratedStarColor(rateValue: number) {
-        const unratedStar = element(by.css(`mat-icon[id="adf-grey-star-${rateValue}"]`));
-        BrowserVisibility.waitUntilElementIsVisible(unratedStar);
-        return unratedStar.getWebElement().getCssValue('color');
     }
 }
